- Made Hint compatible with wxWindows 0.8 - Added some menus
Have to add pictures another day. Would be nice.

There is a problem of which i do not have a solution: if an expression is entered with infinite output, it overloads the event queue (and due to the heaver number of addtions, the text area as well), making it impossible to terminate the process.